requirements.txt
Django==1.8.7
MySQL-python==1.2.5
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-tastypie==0.12.2
django-toolbelt==0.0.1
gunicorn==19.4.1
psycopg2==2.6.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
requests==2.8.1
six==1.10.0
static3==0.6.1
wsgiref==0.1.2
